+++ NEWS.md
@@ -13,6 +13,8 @@
   - Extract resolution from pdb;
   - Renaming arg CRYST1 to CRYSTAL;
 - Function is.crystal: as replacement to is.cryst1;
+- Function visualize.coords: starting some refactoring / optimization;
+- Function toSymbols: started refactoring;
 
 
 # Rpdb 2.3.4